moves event creation content from package event to events client
moves DtAttachRules from config package to events client
