+++ internal/deployment/test_triggered_event_handler.go
@@ -1,19 +1,17 @@
 package deployment
 
 import (
-	"github.com/keptn-contrib/dynatrace-service/internal/config"
 	"github.com/keptn-contrib/dynatrace-service/internal/dynatrace"
-	"github.com/keptn-contrib/dynatrace-service/internal/event"
 )
 
 type TestTriggeredEventHandler struct {
 	event       TestTriggeredAdapterInterface
 	client      dynatrace.ClientInterface
-	attachRules *config.DtAttachRules
+	attachRules *dynatrace.DtAttachRules
 }
 
 // NewTestTriggeredEventHandler creates a new TestTriggeredEventHandler
-func NewTestTriggeredEventHandler(event TestTriggeredAdapterInterface, client dynatrace.ClientInterface, attachRules *config.DtAttachRules) *TestTriggeredEventHandler {
+func NewTestTriggeredEventHandler(event TestTriggeredAdapterInterface, client dynatrace.ClientInterface, attachRules *dynatrace.DtAttachRules) *TestTriggeredEventHandler {
 	return &TestTriggeredEventHandler{
 		event:       event,
 		client:      client,
@@ -24,7 +22,7 @@ func NewTestTriggeredEventHandler(event TestTriggeredAdapterInterface, client dy
 // HandleEvent handles an action finished event
 func (eh *TestTriggeredEventHandler) HandleEvent() error {
 	// Send Annotation Event
-	ie := event.CreateAnnotationEvent(eh.event, eh.attachRules)
+	ie := dynatrace.CreateAnnotationEventDTO(eh.event, eh.attachRules)
 	if ie.AnnotationType == "" {
 		ie.AnnotationType = "Start Tests: " + eh.event.GetTestStrategy()
 	}
@@ -32,7 +30,7 @@ func (eh *TestTriggeredEventHandler) HandleEvent() error {
 		ie.AnnotationDescription = "Start running tests: " + eh.event.GetTestStrategy() + " against " + eh.event.GetService()
 	}
 
-	dynatrace.NewEventsClient(eh.client).SendEvent(ie)
+	dynatrace.NewEventsClient(eh.client).AddAnnotationEvent(ie)
 
 	return nil
 }
