+++ internal/problem/action_triggered_event_handler.go
@@ -4,20 +4,18 @@ import (
 	"errors"
 	"fmt"
 	"github.com/keptn-contrib/dynatrace-service/internal/common"
-	"github.com/keptn-contrib/dynatrace-service/internal/config"
 	"github.com/keptn-contrib/dynatrace-service/internal/dynatrace"
-	"github.com/keptn-contrib/dynatrace-service/internal/event"
 	log "github.com/sirupsen/logrus"
 )
 
 type ActionTriggeredEventHandler struct {
 	event       ActionTriggeredAdapterInterface
 	client      dynatrace.ClientInterface
-	attachRules *config.DtAttachRules
+	attachRules *dynatrace.DtAttachRules
 }
 
 // NewActionTriggeredEventHandler creates a new ActionTriggeredEventHandler
-func NewActionTriggeredEventHandler(event ActionTriggeredAdapterInterface, client dynatrace.ClientInterface, attachRules *config.DtAttachRules) *ActionTriggeredEventHandler {
+func NewActionTriggeredEventHandler(event ActionTriggeredAdapterInterface, client dynatrace.ClientInterface, attachRules *dynatrace.DtAttachRules) *ActionTriggeredEventHandler {
 	return &ActionTriggeredEventHandler{
 		event:       event,
 		client:      client,
@@ -45,11 +43,11 @@ func (eh *ActionTriggeredEventHandler) HandleEvent() error {
 
 	// https://github.com/keptn-contrib/dynatrace-service/issues/174
 	// In addition to the problem comment, send Info and Configuration Change Event to the entities in Dynatrace to indicate that remediation actions have been executed
-	dtInfoEvent := event.CreateInfoEvent(eh.event, eh.attachRules)
+	dtInfoEvent := dynatrace.CreateInfoEventDTO(eh.event, eh.attachRules)
 	dtInfoEvent.Title = "Keptn Remediation Action Triggered"
 	dtInfoEvent.Description = eh.event.GetAction()
 
-	dynatrace.NewEventsClient(eh.client).SendEvent(dtInfoEvent)
+	dynatrace.NewEventsClient(eh.client).AddInfoEvent(dtInfoEvent)
 
 	// this is posting the Event on the problem as a comment
 	comment = fmt.Sprintf("[Keptn triggered action](%s) %s", eh.event.GetLabels()[common.KEPTNSBRIDGE_LABEL], eh.event.GetAction())
